+++ cx-portal/src/types/Config.tsx
@@ -65,6 +65,7 @@ import IDPManagement from 'components/pages/IDPManagement'
 import IDPDetail from 'components/pages/IDPDetail'
 import AppReleaseProcessForm from 'components/pages/AppReleaseProcess/components'
 import CompanyRoles from 'components/pages/CompanyRoles'
+import Deactivate from 'components/pages/AppOverview/Deactivate'
 
 /**
  * ALL_PAGES
@@ -329,6 +330,19 @@ export const ALL_PAGES: IPage[] = [
   { name: PAGES.INTRODUCTION_CONFIRMITY_BODY, element: <CompanyRoles /> },
   { name: PAGES.INTRODUCTION_PARTICIPANT, element: <CompanyRoles /> },
   { name: PAGES.INTRODUCTION_SERVICE_PROVIDER, element: <CompanyRoles /> },
+  {
+    name: PAGES.DEACTIVATE,
+    isRoute: true,
+    element: (
+      <Route
+        key={PAGES.DEACTIVATE}
+        path={PAGES.DEACTIVATE}
+        element={<Deactivate />}
+      >
+        <Route path=":appId" element={<Deactivate />} />
+      </Route>
+    ),
+  },
 ]
 
 export const ALL_OVERLAYS: IOverlay[] = [